Reworked parts of the database. Delete database, remake it, and use 'update-database' to get up to date. Added more DateTime reworks (by Jesus Paz)

+++ Mhotivo.Data/Entities/Notification.cs
@@ -12,12 +12,12 @@ namespace Mhotivo.Data.Entities
         public long Id { get; set; }
         public string NotificationName { get; set; }
         public NotificationType NotificationType { get; set; }
-        public int IdGradeAreaUserGeneralSelected { get; set; }
-        public int GradeIdifNotificationTypePersonal { get; set; }
-        public int StudentId { get; set; }
+        public long IdGradeAreaUserGeneralSelected { get; set; }
+        public long GradeIdifNotificationTypePersonal { get; set; }
+        public long StudentId { get; set; }
         public bool SendingEmail { get; set; }
         public User NotificationCreator { get; set; }
-        public int UserCreatorId { get; set; }
+        public long UserCreatorId { get; set; }
         public string UserCreatorName { get; set; }
         public string Message { get; set; }
         public DateTime Created { get; set; }
