+++ Mhotivo/Models/EnrollModel.cs
@@ -4,7 +4,7 @@ namespace Mhotivo.Models
 {
     public class DisplayEnrollStudents
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [Display(Name = "Nombre Completo")]
         public string FullName { get; set; }
@@ -29,9 +29,9 @@ namespace Mhotivo.Models
     public class EnrollRegisterModel
     {
         [Display(Name = "Grado")]
-        public int GradeId { get; set; }
+        public long GradeId { get; set; }
 
         [Display(Name = "Estudiante")]
-        public int Id { get; set; }
+        public long Id { get; set; }
     }
 }
